backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "amqp==5.2.0",
    "asgiref==3.8.1",
    "billiard==4.2.1",
    "celery==5.3.6",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "click==8.1.7",
    "django-cors-headers==4.3.1",
    "django-debug-toolbar==4.4.6",
    "django-filter==23.5",
    "django==5.0.1",
    "djangorestframework==3.14.0",
    "drf-yasg==1.21.7",
    "gunicorn==21.2.0",
    "inflection==0.5.1",
    "iniconfig==2.0.0",
    "kombu==5.4.2",
    "packaging==24.1",
    "pluggy==1.5.0",
    "prompt-toolkit==3.0.48",
    "psycopg2-binary==2.9.9",
    "pytest-django==4.7.0",
    "pytest==7.4.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.0",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "redis==5.0.1",
    "setuptools>=75.3.0",
    "six==1.16.0",
    "sqlparse==0.5.1",
    "tzdata==2024.2",
    "uritemplate==4.1.1",
    "vine==5.1.0",
    "wcwidth==0.2.13",
]
